Added sandstorm event in world scene

+++ Application/Source/Application.h
@@ -17,8 +17,8 @@ public:
 	void Exit();
 	static bool IsKeyPressed(unsigned short key);
 
-	static inline double getWindowWidth() { return windowWidth; }
-	static inline double getWindowHeight() { return windowHeight; }
+	static inline int getWindowWidth() { return windowWidth; }
+	static inline int getWindowHeight() { return windowHeight; }
 
 	static inline void setWindowSize(int w, int h) { windowWidth = w; windowHeight = h; }
 
@@ -30,7 +30,7 @@ public:
 private:
 
 	//Declare a window object
-	static double windowWidth, windowHeight;
+	static int windowWidth, windowHeight;
 };
 
 #endif
