+++ Application/Source/MGPlayer.cpp
@@ -83,8 +83,8 @@ void MGPlayer::MGPlayerMovements(double dt)
 			}
 		}
 
-		pos_.x += (dir_.x * dt * moveSpeed);
-		pos_.z += (dir_.z * dt * moveSpeed);
+		pos_.x += (dir_.x * (float)dt * moveSpeed);
+		pos_.z += (dir_.z * (float)dt * moveSpeed);
 
 		if (isDPressed && !wasDPressed)
 		{
